[opengl] Refactor to remove the global `no_gc`
